requirements.txt
asgiref
certifi
charset-normalizer
dj-database-url
django
django-environ
gunicorn
idna
psycopg2-binary
pytz
requests
six
sqlparse
typed-ast
urllib3
whitenoise
wrapt
pymysql
python-dateutil